% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar.R
\name{create_calendar}
\alias{create_calendar}
\title{Create a 4k weeks calendar!}
\usage{
create_calendar(birthdate)
}
\arguments{
\item{birthdate}{your birthdate, format: yyyy-mm-dd}
}
\value{
4k weeks calendar as ggplot object
}
\description{
Create a 4k weeks calendar!
}
