% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar.R
\name{create_calendar}
\alias{create_calendar}
\title{Create a 4k weeks calendar!}
\usage{
create_calendar(birthdate, col_past, col_future)
}
\arguments{
\item{birthdate}{your birthdate}

\item{col_past}{rgb color for passed weeks}

\item{col_future}{rgb color for future weeks}
}
\value{
4k weeks calendar as ggplot object
}
\description{
Create a 4k weeks calendar!
}
